requirements-dev.txt
# Development dependencies for AI Brain Python
-r requirements.txt

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
mongomock-motor>=0.0.21

# Code quality
black>=23.7.0
isort>=5.12.0
mypy>=1.5.0
flake8>=6.0.0
pre-commit>=3.3.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0
myst-parser>=2.0.0
sphinx-autodoc-typehints>=1.24.0

# Framework dependencies for development
crewai>=0.1.0
pydantic-ai>=0.1.0
agno>=0.1.0
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.1.0
